requirements.txt
Django==2.0
Pillow==3.3.0
Whoosh==2.7.4
django-extensions==2.2.4
django-filebrowser==3.10.2
django-grappelli==2.11.2
django-haystack==2.8.1
django-simple-captcha==0.5.12
django-tinymce==2.9.0
factory-boy==2.7.0
ipython==5.0.0
mock==2.0.0
psycopg2-binary==2.8.6
pytz==2016.6.1
six==1.10.0
